Crate sp_messenger

Primitives for Messenger.

Modules

endpoint

messages

Structs

ChannelNonce

InherentDataProvider

Provides the set code inherent data.

InherentType

The type of the messenger inherent data.

Enums

InherentError

Inherent specific errors

XdmId

Represent a union of XDM types with their message ID

Constants

INHERENT_IDENTIFIER

Messenger inherent identifier.

MAX_FUTURE_ALLOWED_NONCES

Maximum number of XDMs per domain/channel with future nonces that are allowed to be validated. Any XDM comes with a nonce above Maximum future nonce will be rejected.

Traits

DomainRegistration

Trait to check if the domain is registered.

MessengerApi

Api to provide XDM extraction from Runtime Calls.

NoteChainTransfer

Trait to note cross chain transfer

OnXDMRewards

Trait to handle XDM rewards.

RelayerApi

Api useful for relayers to fetch messages and submit transactions.

StorageKeys

Trait that return various storage keys for storages on Consensus chain and domains