sp_messenger

Module endpoint

Structs

CollectedFee

Fee collected on src_chain for execution of XDM on both the src and dst chains.

EndpointRequest

Request sent by src_endpoint to dst_endpoint.

EndpointRequestWithCollectedFee

Request sent by src_endpoint to dst_endpoint with collected Fee

Enums

Endpoint

Endpoint as defined in the formal spec. Endpoint is an application that can send and receive messages from other chains.

Traits

EndpointHandler

Handler to

Sender

Sender provides abstraction on sending messages to other chains.

Type Aliases

EndpointId

Represents a particular endpoint in a given Execution environment.

EndpointPayload

Endpoint request or response payload.

EndpointResponse

Response for the message request.