Trait sp_messenger::endpoint::EndpointHandler

pub trait EndpointHandler<MessageId> {
    // Required methods
    fn message(
        &self,
        src_domain_id: DomainId,
        message_id: MessageId,
        req: EndpointRequest
    ) -> EndpointResponse;
    fn message_weight(&self) -> Weight;
    fn message_response(
        &self,
        dst_domain_id: DomainId,
        message_id: MessageId,
        req: EndpointRequest,
        resp: EndpointResponse
    ) -> DispatchResult;
    fn message_response_weight(&self) -> Weight;
}

Handler to

handle message request from other domains.

handle requested message responses from other domains.

Required Methods

fn message( &self, src_domain_id: DomainId, message_id: MessageId, req: EndpointRequest ) -> EndpointResponse

Triggered by pallet-messenger when a new inbox message is received and bound for this handler.

fn message_weight(&self) -> Weight

Return the maximal possible consume weight of message

fn message_response( &self, dst_domain_id: DomainId, message_id: MessageId, req: EndpointRequest, resp: EndpointResponse ) -> DispatchResult

Triggered by pallet-messenger when a response for a request is received from dst_domain_id.

fn message_response_weight(&self) -> Weight

Return the maximal possible consume weight of message_response