Trait sp_messenger::DomainTracker

pub trait DomainTracker<StateRoot> {
    fn system_domain_state_roots() -> Vec<StateRoot>;
    fn domain_state_root_storage_key(domain_id: DomainId) -> StorageKey;
}

Implemented by domain registry on system domain or system domain tracker on core domains. This trait supports utilities to verify the message coming from src_domain to system domain. If the message is sent to another core domain, then dst_domain can use this trait and verify the message using System domain as trusted third party.

Required Methods

fn system_domain_state_roots() -> Vec<StateRoot>

Returns a list of state roots of system domain.

fn domain_state_root_storage_key(domain_id: DomainId) -> StorageKey

Returns the storage key that maps to the latest state root of the domain.