Trait sp_messenger::RelayerApi

pub trait RelayerApi<Block: BlockT, BlockNumber, CHash>: Core<Block>
where
    BlockNumber: Encode + Decode,
    CHash: Encode + Decode,
{
    // Provided methods
    fn block_messages(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<BlockMessagesWithStorageKey, ApiError> { ... }
    fn outbox_message_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        msg: CrossDomainMessage<CHash, H256>
    ) -> Result<Option<Block::Extrinsic>, ApiError> { ... }
    fn inbox_response_message_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        msg: CrossDomainMessage<CHash, H256>
    ) -> Result<Option<Block::Extrinsic>, ApiError> { ... }
    fn should_relay_outbox_message(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        dst_chain_id: ChainId,
        msg_id: MessageId
    ) -> Result<bool, ApiError> { ... }
    fn should_relay_inbox_message_response(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        dst_chain_id: ChainId,
        msg_id: MessageId
    ) -> Result<bool, ApiError> { ... }
}

Api useful for relayers to fetch messages and submit transactions.

Provided Methods

fn block_messages( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<BlockMessagesWithStorageKey, ApiError>

Returns all the outbox and inbox responses to deliver. Storage key is used to generate the storage proof for the message.

fn outbox_message_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, msg: CrossDomainMessage<CHash, H256> ) -> Result<Option<Block::Extrinsic>, ApiError>

Constructs an outbox message to the dst_chain as an unsigned extrinsic.

fn inbox_response_message_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, msg: CrossDomainMessage<CHash, H256> ) -> Result<Option<Block::Extrinsic>, ApiError>

Constructs an inbox response message to the dst_chain as an unsigned extrinsic.

fn should_relay_outbox_message( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, dst_chain_id: ChainId, msg_id: MessageId ) -> Result<bool, ApiError>

Returns true if the outbox message is ready to be relayed to dst_chain.

fn should_relay_inbox_message_response( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, dst_chain_id: ChainId, msg_id: MessageId ) -> Result<bool, ApiError>

Returns true if the inbox message response is ready to be relayed to dst_chain.