Trait sp_messenger::StorageKeys

pub trait StorageKeys {
    // Required methods
    fn confirmed_domain_block_storage_key(
        domain_id: DomainId
    ) -> Option<Vec<u8>>;
    fn outbox_storage_key(
        chain_id: ChainId,
        message_key: MessageKey
    ) -> Option<Vec<u8>>;
    fn inbox_responses_storage_key(
        chain_id: ChainId,
        message_key: MessageKey
    ) -> Option<Vec<u8>>;
}

Trait that return various storage keys for storages on Consensus chain and domains

Required Methods

fn confirmed_domain_block_storage_key(domain_id: DomainId) -> Option<Vec<u8>>

Returns the storage key for confirmed domain block on conensus chain

fn outbox_storage_key( chain_id: ChainId, message_key: MessageKey ) -> Option<Vec<u8>>

Returns the outbox storage key for given chain.

fn inbox_responses_storage_key( chain_id: ChainId, message_key: MessageKey ) -> Option<Vec<u8>>

Returns the inbox responses storage key for given chain.

Object Safety

This trait is not object safe.