Crate sp_messenger

Primitives for Messenger.

Modules

endpoint

messages

Structs

InherentDataProvider

Provides the set code inherent data.

InherentType

The type of the messenger inherent data.

Enums

InherentError

Inherent specific errors

Constants

INHERENT_IDENTIFIER

Messenger inherent identifier.

Traits

DomainRegistration

Trait to check if the domain is registered.

MessengerApi

Api to provide XDM extraction from Runtime Calls.

OnXDMRewards

Trait to handle XDM rewards.

RelayerApi

Api useful for relayers to fetch messages and submit transactions.

StorageKeys

Trait that return various storage keys for storages on Consensus chain and domains