sp_messenger::endpoint

Trait EndpointHandler

pub trait EndpointHandler<MessageId> {
    // Required methods
    fn message(
        &self,
        src_chain_id: ChainId,
        message_id: MessageId,
        req: EndpointRequest,
        pre_check_result: DispatchResult,
    ) -> EndpointResponse;
    fn message_weight(&self) -> Weight;
    fn message_response(
        &self,
        dst_chain_id: ChainId,
        message_id: MessageId,
        req: EndpointRequest,
        resp: EndpointResponse,
    ) -> DispatchResult;
    fn message_response_weight(&self) -> Weight;
}

Handler to

handle message request from other chains.

handle requested message responses from other chains.

Required Methods

fn message( &self, src_chain_id: ChainId, message_id: MessageId, req: EndpointRequest, pre_check_result: DispatchResult, ) -> EndpointResponse

Triggered by pallet-messenger when a new inbox message is received and bound for this handler.

fn message_weight(&self) -> Weight

Return the maximal possible consume weight of message

fn message_response( &self, dst_chain_id: ChainId, message_id: MessageId, req: EndpointRequest, resp: EndpointResponse, ) -> DispatchResult

Triggered by pallet-messenger when a response for a request is received from dst_chain_id.

fn message_response_weight(&self) -> Weight

Return the maximal possible consume weight of message_response