sp_messenger

Module messages

Structs

BlockMessagesQuery

A query to fetch block messages for Outbox and Inbox Responses

Channel

Channel describes a bridge to exchange messages between two chains.

ChannelOpenParamsV1

Channel V1 open parameters

CrossDomainMessage

Cross Domain message contains Message and its proof on src_chain.

Message

Message contains information to be sent to or received from another chain.

MessageNonceWithStorageKey

Message with storage key to generate storage proof using the backend.

MessagesWithStorageKey

Set of messages with storage keys to be relayed in a given block.

Enums

ChainId

Identifier of a chain.

ChannelState

State of a channel.

ChannelStateWithNonce

State of channel and nonces when channel is closed.

MessageWeightTag

Message weight tag used to indicate the consumed weight when handling the message

PayloadV1

Payload v1 of the message

Proof

ProtocolMessageRequest

Defines protocol requests performed on chains.

RequestResponse

Protocol message that encompasses request or its response.

VersionedPayload

Versioned message payload

Type Aliases

ChannelId

Channel identity.

MessageId

Unique Id of a message between two chains.

MessageKey

Unique message key for Outbox and Inbox responses

Nonce

Nonce used as an identifier and ordering of messages within a channel. Nonce is always increasing.

ProtocolMessageResponse

Defines protocol requests performed on chains.