Module sp_messenger::messages

Structs

BlockMessageWithStorageKey

Message with storage key to generate storage proof using the backend.

BlockMessagesWithStorageKey

Set of messages with storage keys to be relayed in a given block..

ChannelOpenParams

Channel open parameters

CrossDomainMessage

Cross Domain message contains Message and its proof on src_chain.

FeeModel

Fee model to send a request and receive a response from another chain.

Message

Message contains information to be sent to or received from another chain.

Enums

ChainId

Identifier of a chain.

MessageWeightTag

Message weight tag used to indicate the consumed weight when handling the message

Payload

Payload of the message

Proof

ProtocolMessageRequest

Defines protocol requests performed on chains.

RequestResponse

Protocol message that encompasses request or its response.

VersionedPayload

Versioned message payload

Type Aliases

ChannelId

Channel identity.

MessageId

Unique Id of a message between two chains.

MessageKey

Unique message key for Outbox and Inbox responses

Nonce

Nonce used as an identifier and ordering of messages within a channel. Nonce is always increasing.

ProtocolMessageResponse

Defines protocol requests performed on chains.