sp_messenger

Trait RelayerApi

pub trait RelayerApi<Block: BlockT, BlockNumber, CNumber, CHash>: Core<Block>
where
    BlockNumber: Encode + Decode,
    CNumber: Encode + Decode,
    CHash: Encode + Decode,
{
    // Provided methods
    fn outbox_message_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        msg: CrossDomainMessage<CNumber, CHash, H256>,
    ) -> Result<Option<Block::Extrinsic>, ApiError> { ... }
    fn inbox_response_message_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        msg: CrossDomainMessage<CNumber, CHash, H256>,
    ) -> Result<Option<Block::Extrinsic>, ApiError> { ... }
    fn updated_channels(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<BTreeSet<(ChainId, ChannelId)>, ApiError> { ... }
    fn channel_storage_key(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        chain_id: ChainId,
        channel_id: ChannelId,
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn open_channels(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<BTreeSet<(ChainId, ChannelId)>, ApiError> { ... }
    fn block_messages_with_query(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        query: BlockMessagesQuery,
    ) -> Result<MessagesWithStorageKey, ApiError> { ... }
    fn channels_and_state(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
    ) -> Result<Vec<(ChainId, ChannelId, ChannelStateWithNonce)>, ApiError> { ... }
    fn first_outbox_message_nonce_to_relay(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        dst_chain_id: ChainId,
        channel_id: ChannelId,
        from_nonce: Nonce,
    ) -> Result<Option<Nonce>, ApiError> { ... }
    fn first_inbox_message_response_nonce_to_relay(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        dst_chain_id: ChainId,
        channel_id: ChannelId,
        from_nonce: Nonce,
    ) -> Result<Option<Nonce>, ApiError> { ... }
}

Api useful for relayers to fetch messages and submit transactions.

Provided Methods

fn outbox_message_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, msg: CrossDomainMessage<CNumber, CHash, H256>, ) -> Result<Option<Block::Extrinsic>, ApiError>

Constructs an outbox message to the dst_chain as an unsigned extrinsic.

fn inbox_response_message_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, msg: CrossDomainMessage<CNumber, CHash, H256>, ) -> Result<Option<Block::Extrinsic>, ApiError>

Constructs an inbox response message to the dst_chain as an unsigned extrinsic.

fn updated_channels( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<BTreeSet<(ChainId, ChannelId)>, ApiError>

Returns the list of channels updated in the given block.

fn channel_storage_key( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, chain_id: ChainId, channel_id: ChannelId, ) -> Result<Vec<u8>, ApiError>

Returns storage key for channels for given chain and channel id.

fn open_channels( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<BTreeSet<(ChainId, ChannelId)>, ApiError>

Returns all the open channels to other chains.

fn block_messages_with_query( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, query: BlockMessagesQuery, ) -> Result<MessagesWithStorageKey, ApiError>

Returns outbox and inbox responses from given nonce to maximum allowed nonce per block Storage key is used to generate the storage proof for the message.

fn channels_and_state( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ) -> Result<Vec<(ChainId, ChannelId, ChannelStateWithNonce)>, ApiError>

Returns all the channels to other chains and their local Channel state.

fn first_outbox_message_nonce_to_relay( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, dst_chain_id: ChainId, channel_id: ChannelId, from_nonce: Nonce, ) -> Result<Option<Nonce>, ApiError>

Returns the first outbox message nonce that should be relayed to the dst_chain.

fn first_inbox_message_response_nonce_to_relay( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, dst_chain_id: ChainId, channel_id: ChannelId, from_nonce: Nonce, ) -> Result<Option<Nonce>, ApiError>

Returns the first inbox response message nonce that should be relayed to the dst_chain.